requirements.txt
bootstrap-admin==0.4.0
certifi==2018.11.29
chardet==3.0.4
Django==2.1
django-js-asset==1.1.0
django-mdeditor==0.1.9
django-pure-pagination==0.3.0
idna==2.8
Markdown==3.0.1
mysqlclient==1.3.14
Pillow==5.3.0
Pygments==2.3.0
pytz==2018.7
requests==2.21.0
urllib3==1.24.1
